+++ src/main/java/org/silentsoft/stopwatch/NumberFormatter.java
@@ -4,8 +4,19 @@ import java.text.NumberFormat;
 
 public final class NumberFormatter {
 
+    private static NumberFormat percentageFormat;
     private static NumberFormat millisecondsFormat;
     private static NumberFormat secondsFormat;
+
+    public static String percentage(double number) {
+        if (percentageFormat == null) {
+            percentageFormat = NumberFormat.getInstance();
+            percentageFormat.setMinimumFractionDigits(1);
+            percentageFormat.setMaximumFractionDigits(1);
+        }
+
+        return percentageFormat.format(number).concat("%");
+    }
 
     public static String milliseconds(double number) {
         if (millisecondsFormat == null) {
@@ -19,6 +30,7 @@ public final class NumberFormatter {
         if (secondsFormat == null) {
             secondsFormat = NumberFormat.getInstance();
             secondsFormat.setMinimumFractionDigits(3);
+            secondsFormat.setMaximumFractionDigits(3);
         }
 
         return secondsFormat.format(number).concat("s");